+++ src/App.svelte
@@ -1,37 +1,5 @@
 <script>
-  // Life Cycle
-  import { beforeUpdate } from "svelte";
-
-  // Components
-  import AppState from "UI/AppState.svelte";
-  import WelcomeScreen from "./WelcomeScreen.svelte";
-  import LessonsScreen from "./LessonsScreen.svelte";
-
-  // Stores
-  import { APP_STATE, IS_DEBUG } from "Stores/AppState.js";
-  import { USER_OBJ } from "Stores/UserState.js";
-
-  // Helpers and Enums
-  import { setLsItem } from "Scripts/LocalStorageHelper.js";
-  import { LSKeyEnums, AppStateEnums } from "Scripts/enum.js";
-
-  // Reactive variables
-  $: appState = $APP_STATE.state;
-
-  const rightNow = Date.now();
-  // Setting Threshold to 18 hours.
-  const resetThreshold = 1000 * 60 * 60 * 18;
-
-  function changeAppState({ detail }) {
-    APP_STATE.setState(detail);
-  }
-
-  beforeUpdate(() => {
-    if (rightNow - $USER_OBJ.active_time > resetThreshold) {
-      localStorage.clear();
-      window.location.reload();
-    }
-  });
+  import Nav from "Global/Nav.svelte";
 </script>
 
 <style lang="scss" global>
@@ -39,16 +7,14 @@
 </style>
 
 <main>
-  {#if $IS_DEBUG}
-    <AppState />
-  {/if}
+  <Nav />
 
-  {#if !appState || appState == AppStateEnums.appStart}
-    <WelcomeScreen on:updateState={changeAppState} />
-  {/if}
-
-  {#if appState == AppStateEnums.lessonSelect}
-    <LessonsScreen on:changeDifficulty={changeAppState} />
-  {/if}
-
+  <div class="container">
+    <header>
+      <h1>
+        <small>Welcome to</small>
+        Typer
+      </h1>
+    </header>
+  </div>
 </main>
