+++ src/components/Global/Nav.svelte
@@ -1,0 +1,5 @@
+<nav class="nav">
+  <a href="/" class="nav__link -home">Home</a>
+  <a href="/" class="nav__link -code">Code</a>
+  <a href="/" class="nav__link -type">Type</a>
+</nav>
