+++ src/styles/abstracts/_base-vars.scss
@@ -1,3 +1,9 @@
+// Links
+$text-color: #f5f5f5;
+$link-color: #3b8332; //#ff5c62;
+$link-border-color: #e2fff3; //#ffe2e3;
+$link-bg-color: #e5ebe8; //#fff4f5;
+
 $electricLime: #bada55;
 
 $mandalay: #b4891d;
