+++ src/styles/abstracts/_break-points.scss
@@ -4,5 +4,5 @@ $grid-breakpoints: (
   md: 768px,
   lg: 1024px,
   xl: 1274px,
-  xxl: 1548px,
+  xxl: 1548px
 );
