+++ src/styles/abstracts/_misc.scss
@@ -10,6 +10,6 @@ $layout: (
     sm: $padding - 0.25,
     base: $padding,
     lg: $padding + 0.5,
-    xl: $padding + 0.75,
-  ),
+    xl: $padding + 0.75
+  )
 );
