--- src/styles/base/_base.scss
@@ -1,64 +1,0 @@
-@import url("https://fonts.googleapis.com/css?family=Roboto+Mono:400,500,700&display=swap");
-
-*,
-*:before,
-*:after {
-  box-sizing: border-box;
-}
-
-body {
-  margin: 0;
-  overflow: hidden;
-  height: 100vh;
-  width: 100vw;
-  font-family: "Roboto Mono", serif;
-}
-
-p {
-  margin: 0;
-}
-
-a {
-  transition: all 0.25s ease;
-  text-decoration: none;
-}
-
-main {
-  overflow: hidden;
-  height: 100vh;
-  width: 100vw;
-}
-
-.app-state {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: $padding/2;
-  position: fixed;
-  width: 100%;
-  bottom: 0;
-  left: 0;
-  background: rgba($accent, 0.4);
-  text-align: right;
-  z-index: 100;
-}
-
-.container {
-  width: 100%;
-  height: 100%;
-  color: $white;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  @include gradientBG;
-}
-
-.inner {
-  width: 75%;
-  max-width: 1000px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
