--- src/styles/base/_codeDisplay.scss
@@ -1,69 +1,0 @@
-.code-display {
-  position: absolute;
-  right: 0;
-  bottom: 0;
-  width: 60vw;
-  height: 50vh;
-  z-index: 0;
-  transition: all 0.25s ease;
-
-  &.hidden {
-    bottom: -45%;
-  }
-}
-
-.tabs {
-  display: flex;
-  justify-content: space-between;
-}
-
-.tab-group {
-  display: flex;
-}
-
-.tab {
-  cursor: pointer;
-  background: $white;
-  padding: $padding/2;
-  border: 1px solid $border-color-light;
-  border-top: 4px solid $border-color-light;
-
-  &.focus {
-    border-top: 4px solid orangered;
-  }
-
-  &:focus {
-    outline: none;
-  }
-}
-
-.displays {
-  background: $bg-light;
-  padding: $padding/2;
-  height: calc(100% - 37px);
-  position: relative;
-}
-
-.display {
-  transition: all 0.25s ease;
-  opacity: 0;
-  position: absolute;
-  top: 0;
-  left: 0;
-  margin: 0;
-  width: 100%;
-  height: 100%;
-  overflow-x: auto;
-
-  &.focus {
-    opacity: 1;
-  }
-}
-
-.display-line {
-  padding: 4px $padding/2;
-
-  &:nth-child(even) {
-    background: $white;
-  }
-}
