--- src/styles/base/_index.scss
@@ -1,3 +1,0 @@
-@import "base";
-@import "buttons";
-@import "codeDisplay";
