+++ src/styles/components/_index.scss
@@ -1,3 +1,1 @@
-@import "lesson";
-@import "loading";
-@import "render";
+@import "nav";
