--- src/styles/components/_lesson.scss
@@ -1,96 +1,0 @@
-.lessons-select {
-  margin-top: $padding;
-}
-
-.lesson-wrapper {
-  font-family: "Roboto Mono", monospace;
-  position: relative;
-  padding: 0;
-  @include gradientBG();
-  border: 5px solid $white;
-  width: 40vw;
-  height: 100vh;
-  box-shadow: 4px 0 4px rgba($black, 0.3);
-  z-index: 1;
-}
-
-// Code - Block
-.code {
-  color: $text-color-light;
-}
-
-.row {
-  min-height: 25px;
-  display: flex;
-  flex-wrap: wrap;
-  padding: 0 $padding/2;
-}
-
-.character {
-  padding: 0;
-  min-width: 7px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-// Code - User Interface
-.cursor {
-  background: $accent;
-  animation: blink 1.25s ease infinite;
-}
-
-.complete,
-.correct {
-  color: $correct;
-  border-bottom: 1px solid $correct;
-  font-weight: bold;
-}
-.completeRow,
-.currentRow {
-  background: $complete;
-}
-
-// Lesson Description
-.lesson {
-  background: $white;
-  padding: 1rem;
-  position: absolute;
-  bottom: 0;
-  max-height: 50vh;
-  width: 100%;
-  overflow-y: auto;
-}
-
-.lesson__title {
-  margin: 0;
-  font-size: 1.25rem;
-}
-
-// Lesson List
-.lesson-list {
-  list-style: none;
-  margin: $padding / 2 0;
-  padding: 0;
-}
-
-.lesson-item {
-  &:not(:last-child) {
-    margin-bottom: $padding / 2;
-  }
-}
-
-.lesson-link {
-  display: inline-block;
-  padding: $padding/2 $padding;
-  background: $brand-secondary;
-  color: $text-color-light;
-
-  .hasCompleted & {
-    background: $accent;
-  }
-
-  &:hover {
-    background: $brand-secondary-dark;
-  }
-}
