--- src/styles/components/_loading.scss
@@ -1,34 +1,0 @@
-.loading {
-  position: relative;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  span {
-    display: inline-block;
-    margin: 0 2px;
-    opacity: 0;
-    min-width: 10px;
-    transform: scale(0.6);
-    transform-origin: center bottom;
-
-    @for $i from 1 to 25 {
-      &:nth-child(#{$i}) {
-        animation: loading 1.5s linear #{$i/35}s infinite alternate;
-      }
-    }
-  }
-}
-
-@keyframes loading {
-  0% {
-    opacity: 0;
-    transform: scale(0.6);
-  }
-  30%,
-  100% {
-    opacity: 1;
-    transform: scale(1.2);
-  }
-}
