--- src/styles/components/_render.scss
@@ -1,8 +1,0 @@
-.render {
-  position: absolute;
-  width: 60vw;
-  height: 100vh;
-  right: 0;
-  top: 0;
-  padding: $padding;
-}
