+++ src/styles/styles.scss
@@ -1,6 +1,6 @@
 @import "abstracts/index";
 @import "helpers/index";
-@import "base/index";
+// @import "base/index";
 @import "components/index";
-@import "helpers/index";
-@import "base/index";
+// @import "helpers/index";
+// @import "base/index";
